ref: create spotify provider folder
Move all spotify related connection stuff to `connection.rs`,
shared function & variables in `mod.rs` (unsure if that's good or not,
maybe should create shared.rs?)
and currently_playing stuff in `playing.rs`.
